% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neMDBD.R
\name{get_P_tsne}
\alias{get_P_tsne}
\title{get_P_tsne}
\usage{
get_P_tsne(X, perplexity)
}
\arguments{
\item{X}{Pre-processed X maatrix}

\item{perplexity}{Perplexity of t-SNE algorithm.}
}
\value{
P matrix, unnormalized P matrix, and beta (1/2sigma^2).
}
\description{
R function to calculate P matrix of t-SNE (https://github.com/jdonaldson/rtsne)
}
\examples{
get_P_tsne(X, perplexity)
}
