% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neMDBD.R
\name{gradient_compute}
\alias{gradient_compute}
\title{gradient_compute}
\usage{
gradient_compute(Y, P)
}
\arguments{
\item{Y}{Matrix. t-SNE embedding.}

\item{P}{Matrix. Similarity matrix in t-SNE.}
}
\value{
Gradients for t-SNE loss function at embedding Y.
}
\description{
computes gradients for t-SNE loss function
}
\examples{
gradient_compute(Y, P)
}
