% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neMDBD.R
\name{gradient_compute_pointwise}
\alias{gradient_compute_pointwise}
\title{gradient_compute_pointwise}
\usage{
gradient_compute_pointwise(yy, Y, P)
}
\arguments{
\item{yy}{Vector. Embedding of the point that we want to calculate loss with.}

\item{Y}{Matrix. Embedding matrix of other points.}

\item{P}{Matrix. Similarity matrix in t-SNE.}
}
\value{
Gradients for t-SNE loss function at embedding yy.
}
\description{
computes gradients for t-SNE loss function pointwise
}
\examples{
gradient_compute_pointwise(yy, Y, P)
}
