% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neMDBD.R
\name{LOO_loss_compute}
\alias{LOO_loss_compute}
\title{LOO_loss_compute}
\usage{
LOO_loss_compute(yy, Y, PMat)
}
\arguments{
\item{yy}{Vector. Proposed embedding point of the point x that we want to calculate loss with.}

\item{Y}{Matrix. Embedding matrix of X.}

\item{PMat}{Matrix. Simalarity matrix (P matrix) of 'rbind(x,X)'.}
}
\value{
The LOO loss for the point yy.
}
\description{
get t-SNE LOO loss for specified point
}
\examples{
LOO_loss_compute(yy, Y, Mat)
}
