% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neMDBD.R
\name{LOO_loss_fast}
\alias{LOO_loss_fast}
\title{LOO_loss_fast}
\usage{
LOO_loss_fast(yy, Y, YDistSqP1, PMat)
}
\arguments{
\item{yy}{Vector. Embedding of the point that we want to calculate loss with.}

\item{Y}{Matrix. Embedding matrix of other points.}

\item{YDistSqP1}{Matrix. Squared distance matrix plus 1 of 'rbind(c(0,0), Y)' for fast computation.}

\item{PMat}{Matrix. Simalarity matrix (P matrix) of all points (with the point we want to calculate loss with as the first point).}
}
\value{
The LOO loss for the point yy.
}
\description{
get t-SNE LOO loss for specified point
}
\examples{
LOO_loss_fast(yy, Y, YDistSqP1, Mat)
}
