% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neMDBD.R
\name{perturbation_score_compute}
\alias{perturbation_score_compute}
\title{perturbation_score_compute}
\usage{
perturbation_score_compute(
  X,
  Y,
  perplexity,
  length,
  approx,
  ind = NULL,
  no.cores = NULL,
  initial_dims = 50,
  pca_center = TRUE,
  pca_scale = FALSE
)
}
\arguments{
\item{X}{Matrix. Original Data.}

\item{Y}{Matrix. Embedding of original data.}

\item{perplexity}{Integer. Perplexity parameter to use, should be the same as when calculating the embedding Y.}

\item{length}{Numeric. Length of perturbation.}

\item{approx}{Integer. 0: no approximation; 1: reuse the original PCA result; 2: with approximation to similarity matrix.}

\item{ind}{Vector. Specify indices of points for perturbation score calculation. (default: NULL (calculate for all points))}

\item{no.cores}{Integer. Number of cores for parallel computing. (default: NULL (no parallel computing))}

\item{initial_dims}{Integer. The number of dimensions that should be retained in the initial PCA step in t-SNE.}

\item{pca_center}{Logical. Should data be centered before pca is applied? (default: TRUE)}

\item{pca_scale}{Logical. Should data be scaled before pca is applied? (default: FALSE)}
}
\value{
Perturbation score for the i-th point in the data.
}
\description{
Given the perturbation direction and length for specified point, calculate the perturbation scores.
}
\examples{
perturbation_score_compute(X, Y, perplexity, length, approx)
}
