% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neMDBD.R
\name{perturbation_score_compute_pointwise}
\alias{perturbation_score_compute_pointwise}
\title{perturbation_score_compute_pointwise}
\usage{
perturbation_score_compute_pointwise(
  i,
  X,
  Y,
  perplexity,
  dir_vec,
  length,
  Ydist_sq = NULL,
  initial_dims = 50,
  PCA_result = NULL,
  P_unnormalized = NULL,
  beta = NULL,
  approx = 0
)
}
\arguments{
\item{i}{Integer. Calculate the perturbation score for the i-th point.}

\item{X}{Matrix. Original Data.}

\item{Y}{Matrix. Embedding of original data.}

\item{perplexity}{Integer. Perplexity parameter to use, should be the same as when calculating the embedding Y.}

\item{dir_vec}{List. A list of perturbation directions.}

\item{length}{Numeric. Length of perturbation.}

\item{Ydist_sq}{Matrix. Distance matrix of embedding Y, can be calculated by as.matrix(dist(Y)**2).}

\item{initial_dims}{Integer. The number of dimensions that should be retained in the initial PCA step in t-SNE.}

\item{PCA_result}{PCA result of the original data. Can be calculated as prcomp(X).}

\item{P_unnormalized}{Matrix. Similarity matrix before normalization.}

\item{beta}{Vector. Vector of 1/(2sigma_i^2) in Gaussian kernel in input space in t-SNE.}

\item{approx}{Integer. 0: no approximation; 1: reuse the original PCA result; 2: with approximation to similarity matrix.}
}
\value{
Perturbation score for the i-th point in the data.
}
\description{
Given the perturbation direction and length for specified point, calculate the perturbation score.
}
\examples{
perturbation_score_compute_pointwise(i, X, Y, perplexity, dir_vec, length, ...)
}
