% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neMDBD.R
\name{singularity_score_compute}
\alias{singularity_score_compute}
\title{singularity_score_compute}
\usage{
singularity_score_compute(Y, P)
}
\arguments{
\item{Y}{Matrix. t-SNE embedding.}

\item{P}{Matrix. Similarity matrix in t-SNE.}
}
\value{
Hessian matrix for t-SNE loss function at embedding Y.
}
\description{
computes the singularity score to diagnose fracture-inducing discontinuity
}
\examples{
singularity_score_compute(Y, P)
}
